# postgres_execution_package/api_spec/openapi.yaml
openapi: 3.0.3
info:
  title: منصة قطع غيار السيارات - Auto Parts Marketplace API
  version: "1.9.0"
  description: >
    مواصفة API تنفيذية، تغطي 15 عملية (Operations) عبر 14 مسارًا (Paths):
    الواجهات التمثيلية الخمس الأصلية الموثّقة بالتفصيل في التصميم التفصيلي
    (الحزمة 2)، بالإضافة إلى 5 عمليات خدمة Auth (اعتماد Contract Extension &
    Implementation Plan): تسجيل خروج، عرض/إضافة/إزالة وسائل الهوية، وعرض
    مزوّدي الهوية المفعَّلين — فتصبح عمليات Auth تحديدًا 6 عمليات شاملة
    /auth/login الأصلية. وبالإضافة أيضًا إلى 5 عمليات خدمة PCT (الكتالوج):
    اقتراح قطعة، عرضها، اعتمادها (مدير النظام فقط، REQ-PCT-002)، إضافة اسم
    مترجَم، وتسجيل رقم OEM. وبالإضافة إلى 7 عمليات خدمة VCT (كتالوج
    السيارات): اقتراح واعتماد شركة مصنّعة (مدير النظام فقط، REQ-VCT-002)،
    عرضها، اقتراح موديل تحت شركة معتمَدة (REQ-VCT-003)، إضافة جيل، إضافة
    فئة سيارة، وعرض فئة سيارة. وبالإضافة إلى 4 عمليات خدمة CMP (التوافق):
    إنشاء سجل توافق بين قطعة معتمَدة وفئة صالحة (مدير النظام فقط،
    REQ-CMP-001/002)، عرض سجل، عرض كل سجلات قطعة، وأرشفته (مدير النظام
    فقط، REQ-CMP-003). وبالإضافة أيضًا إلى خدمتَي Store وInventory: إنشاء/
    عرض متجر، تغيير حالته (مدير نظام أو مشرف، REQ-STR-004)، نقل ملكيته
    (مدير نظام حصريًا، REQ-STR-006)؛ وعرض/تحديث كمية/تسعير/إخفاء/إظهار/
    أرشفة عنصر مخزون (البائع المالك فقط، REQ-STR-019) إلى جانب /inventory-items
    الأصلية (إنشاء العنصر، مع Idempotency-Key مطلوب). جميع مسارات هذا
...
    كل عملية موسومة بامتداد x-traceability
...
  x-governance: >
...
servers:
...
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: >
        جلسة مستخدم صادرة عن وحدة IAM بعد تسجيل دخول ناجح (REQ-IAM-003)،
        مدعومة بجدول iam.sessions (CR-013). تُسلَّم عبر Set-Cookie بخصائص
        HttpOnly (لا يقرأها JavaScript) وSecure (HTTPS فقط) وSameSite=Lax
        كحد أدنى. يُخزَّن على الخادم بصمة SHA-256 للتوكن فقط، لا التوكن
        الخام (اتساقًا مع REQ-SEC-002). غير مطلوبة لعمليات الوصول العام مثل
        تسجيل الدخول نفسه أو البحث أو عرض مزوّدي الهوية المفعَّلين.
  parameters:
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >
        معرّف تتبع اختياري يرفقه المستهلك؛ إن غاب يُنشئ النظام معرّفًا جديدًا
        عند نقطة الدخول (SAD الحزمة 4، DD الحزمة 2 القسم 2.4).
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      description: نموذج الخطأ الموحّد المعتمد لجميع واجهات API (DD الحزمة 2، القسم 2.1/2.2)
      required: [correlation_id, error_code, message]
      properties:
        correlation_id:
          type: string
          format: uuid
        error_code:
          type: string
          example: "PART_NOT_APPROVED"
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
...
    CatalogPartCreateRequest:
      type: object
      required: [category_id]
      properties:
        category_id:
          type: string
          format: uuid
    CatalogPartResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        category_id: { type: string, format: uuid }
        status: { type: string, enum: [proposed, approved, archived] }
    LocalizedNameCreateRequest:
      type: object
      required: [name_value, name_kind]
      properties:
        name_value: { type: string }
        name_kind: { type: string, enum: [canonical, local, english, synonym] }
        locale: { type: string, nullable: true }
    LocalizedNameResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name_value: { type: string }
        name_kind: { type: string }
        locale: { type: string, nullable: true }
    OemNumberCreateRequest:
      type: object
      required: [manufacturer_ref_id, oem_number]
      properties:
        manufacturer_ref_id: { type: string, format: uuid }
        oem_number: { type: string }
    OemNumberResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        manufacturer_ref_id: { type: string, format: uuid }
        oem_number: { type: string }
...
paths:
...
  /api/v1/pct/parts:
    post:
      operationId: proposeCatalogPart
      summary: تقديم طلب إضافة قطعة جديدة للكتالوج
      x-traceability:
        requirements: [REQ-PCT-001, REQ-PCT-007]
        decisions: []
      security:
        - sessionAuth: []
      parameters: [{ $ref: '#/components/parameters/CorrelationIdHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CatalogPartCreateRequest' }
      responses:
        '201':
          description: أُنشئت القطعة بحالة proposed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogPartResponse' }
        '401': { description: لا جلسة صالحة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: category_id غير موجود (FK), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/pct/parts/{partId}:
    get:
      operationId: getCatalogPart
      summary: عرض تفاصيل قطعة كتالوج بمعرّفها
      x-traceability:
        requirements: [REQ-PCT-001, REQ-PCT-002]
        decisions: []
      security:
        - sessionAuth: []
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: partId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: بيانات القطعة
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogPartResponse' }
        '404': { description: غير موجودة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/pct/parts/{partId}/approve:
    post:
      operationId: approveCatalogPart
      summary: اعتماد قطعة كتالوج (proposed → approved)
      x-traceability:
        requirements: [REQ-PCT-002]
        decisions: ["يستوجب تحديد آلية تفويض دور مدير النظام قبل التنفيذ — انظر ملاحظة الصلاحيات"]
      security:
        - sessionAuth: []
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: partId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: أصبحت القطعة معتمَدة
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogPartResponse' }
        '401': { description: لا جلسة صالحة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: لا صلاحية اعتماد (مدير نظام فقط، REQ-PCT-002), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: غير موجودة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: انتقال حالة غير مسموح (InvalidCatalogPartStatusError), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/pct/parts/{partId}/names:
    post:
      operationId: addCatalogPartLocalizedName
      summary: إضافة اسم (قياسي/محلي/إنجليزي/مرادف) لقطعة كتالوج
      x-traceability:
        requirements: [REQ-PCT-003]
        decisions: []
      security:
        - sessionAuth: []
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: partId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocalizedNameCreateRequest' }
      responses:
        '201':
          description: أُضيف الاسم
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LocalizedNameResponse' }
        '400': { description: قيمة اسم فارغة أو نوع اسم غير معروف, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: القطعة غير موجودة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/pct/parts/{partId}/oem-numbers:
    post:
      operationId: addCatalogPartOemNumber
      summary: تسجيل رقم OEM لقطعة كتالوج
      x-traceability:
        requirements: [REQ-PCT-004, REQ-PCT-005]
        decisions: []
      security:
        - sessionAuth: []
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: partId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OemNumberCreateRequest' }
      responses:
        '201':
          description: أُضيف رقم OEM
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OemNumberResponse' }
        '404': { description: القطعة غير موجودة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: 'رقم OEM مكرَّر لنفس الشركة المصنِّعة (REQ-PCT-005، DuplicateOemNumberError)', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...